Removed MapView's constructor, made some methods abstract.
